refactor: move _numpy helper under dspy/utils/
